* saving and loading tweakCount in .tas and fm2 * Added string "Tweak Count" to Replay loading dialog * Tasedit: View - "Show Tweak count" * Tasedit: rewrote project save/load functions * Tasedit: fixed old bug caused by not initializing input ports * Tasedit: checking movie fourscore flag when creating/opening project; only show columns for pads 3 and 4 if fourscore is attached

+++ src/drawing.cpp
@@ -420,7 +420,7 @@ void DrawTextTransWH(uint8 *dest, uint32 width, uint8 *textmsg, uint8 fgcolor, i
 		wid = JoedCharWidth(*textmsg);
 
 		int newx = x+wid;
-		if(newx >= width) { x=beginx; y+=8; }
+		if(newx >= (int)width) { x=beginx; y+=8; }
 
 		for(int ny=0; ny<7; ++ny)
 		{
